core/api/v1/budget_management/filters.py
from ninja import Schema


class CurrencyFilters(Schema):
    search: str | None = None


class BudgetFilters(Schema):
    search: str | None = None


class CategoryFilters(Schema):
    search: str | None = None


class OperationFilters(Schema):
    search: str | None = None
