requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-ninja==1.3.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
sqlparse==0.5.3
typing_extensions==4.12.2
